backend/requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.9.0
blis==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
coverage==7.8.0
cryptography==44.0.0
cymem==2.0.10
daphne==4.1.2
distro==1.9.0
Django==5.1.5
django-cors-headers==4.7.0
django-environ==0.12.0
docx2txt==0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.12
filelock==3.16.1
Flask==3.1.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.2.0
murmurhash==1.0.11
mysqlclient==2.2.7
namex==0.0.8
narwhals==1.19.0
nltk==3.9.1
numpy==2.1.3
openai==1.68.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pafy==0.5.5
pandas==2.2.3
pdfminer.six==20231228
pdfminer3==2018.12.3.0
pdfplumber==0.11.4
pillow==11.1.0
plotly==5.24.1
pluggy==1.5.0
preshed==3.0.9
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyresparser==1.0.6
pyrsistent==0.20.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
service-identity==24.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.0
starlette==0.46.2
streamlit==1.41.1
streamlit-tags==1.2.8
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
thinc==8.3.3
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
Twisted==24.11.0
txaio==23.1.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.2
waitress==3.0.2
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3
zope.interface==7.2
